:green_heart: Extract store file from model file in spec
