+++ spec/goa_model_gen/project1/model/component1_only_store.go
@@ -1,0 +1,1 @@
+package model
